% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete-records.R
\name{dbDeleteAllRecords}
\alias{dbDeleteAllRecords}
\title{delete all records in an existing table}
\usage{
dbDeleteAllRecords(con, name, key.df = NULL, ...)
}
\arguments{
\item{con}{a DBConnection oject}

\item{name}{a character string specifying the name of an existing DBMS table}

\item{key.df}{NULL | a dataframe of key value pairs designating records to delete}
}
\description{
The default `dbDeleteAllRecords()` method performs checks that
the targeted table exists
}
\details{
The sql generation uses the function `sqlDeleteAllRecords()` another S4 generic defined in this package.
Default sql generation is appropriate for MariaDB and MySQL.
}
\seealso{
Other DBIConnection generics: \code{\link{dbAddIndex}},
  \code{\link{dbAddPrimaryKey}}, \code{\link{dbDropTable}},
  \code{\link{dbReplaceInTable}},
  \code{\link{dbUpdateTable}}
}
\concept{DBIConnection generics}
