% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop-table.R
\name{dbDropTable}
\alias{dbDropTable}
\title{Drop an existing table}
\usage{
dbDropTable(con, name, ...)
}
\arguments{
\item{con}{a DBConnection oject}

\item{name}{a character string specifying the name of an existing DBMS table}

\item{primary_key}{a character vector of column names which will form the primary key

    To specify a composite Primary Key for the `symbol` and `date` columns use

    primary_key = c("symbol","date" )

For a table with a Primary Key defined, each records must have a unqiue primary key. Attempts to `INSERT`
records with a duplicate Primary Key to an existing record in the DB will cause an ERROR}
}
\description{
The sql generation uses the function `sqlDropTable()` another S4 generic defined in this package.
Default sql generation is appropriate for MariaDB and MySQL.
}
\seealso{
Other DBIConnection generics: \code{\link{dbAddIndex}},
  \code{\link{dbAddPrimaryKey}},
  \code{\link{dbDeleteAllRecords}},
  \code{\link{dbReplaceInTable}},
  \code{\link{dbUpdateTable}}
}
\concept{DBIConnection generics}
