% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete-records.R, R/drop-table.R, R/hidden.R,
%   R/sql-generation.R, R/table-indexing.R, R/table-replace.R, R/table-update.R
\docType{methods}
\name{dbDeleteAllRecords,DBIConnection-method}
\alias{dbDeleteAllRecords,DBIConnection-method}
\alias{dbDropTable,DBIConnection-method}
\alias{hidden_aliases}
\alias{sqlOnClause,DBIConnection-method}
\alias{sqlSetClause,DBIConnection-method}
\alias{sqlUpdateTable,DBIConnection-method}
\alias{sqlAlterTableWithPrimaryKey,DBIConnection-method}
\alias{sqlAddIndex,DBIConnection-method}
\alias{sqlDeleteAllRecords,DBIConnection-method}
\alias{dbAddPrimaryKey,DBIConnection-method}
\alias{dbAddIndex,DBIConnection-method}
\alias{sqlReplaceInTable,DBIConnection-method}
\alias{dbReplaceInTable,DBIConnection-method}
\alias{dbUpdateTable,DBIConnection-method}
\title{Internal page for hidden aliases}
\usage{
\S4method{dbDeleteAllRecords}{DBIConnection}(con, name, key.df = NULL,
  ...)

\S4method{dbDropTable}{DBIConnection}(con, name, ...)

\S4method{sqlOnClause}{DBIConnection}(con, x, y, by)

\S4method{sqlSetClause}{DBIConnection}(con, x, y, set)

\S4method{sqlUpdateTable}{DBIConnection}(con, x, y, set = NULL,
  by = NULL)

\S4method{sqlAlterTableWithPrimaryKey}{DBIConnection}(conn, target.table,
  pk)

\S4method{sqlAddIndex}{DBIConnection}(conn, table, index, unique = FALSE)

\S4method{sqlDeleteAllRecords}{DBIConnection}(conn, name)

\S4method{dbAddPrimaryKey}{DBIConnection}(con, name, primary_key, ...)

\S4method{dbAddIndex}{DBIConnection}(con, name, index, unique = FALSE,
  ...)

\S4method{sqlReplaceInTable}{DBIConnection}(con, table, values,
  row.names = NA, ...)

\S4method{dbReplaceInTable}{DBIConnection}(conn, name, value, ...,
  row.names = NULL)

\S4method{dbUpdateTable}{DBIConnection}(con, target.table, data.update,
  set = NULL, by = NULL)
}
\description{
For S4 methods that require a documentation entry but only clutter the index.
}
\keyword{internal}
