% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-generation.R
\name{sqlAddIndex}
\alias{sqlAddIndex}
\title{generate sql for an alter table statment adding an index to an existing DBMS table}
\usage{
sqlAddIndex(conn, table, index, unique = FALSE, ...)
}
\arguments{
\item{conn}{a DBIConnector Object}

\item{table}{table in database to which primary key will be added}

\item{index}{a character vector of column names which make up the index for the table}

\item{unqiue}{boolean specifying if index is unique (a rows must have a unique value)}
}
\description{
suggest that this be done prior to adding any data to the table.
}
\details{
`sqlAddIndex()` composes a single SQL statement to add a composite primary key
to a table in the database
}
\seealso{
Other SQL generation: \code{\link{sqlAlterTableWithPrimaryKey}},
  \code{\link{sqlOnClause}},
  \code{\link{sqlReplaceInTable}},
  \code{\link{sqlSetClause}}, \code{\link{sqlUpdateTable}}
}
\concept{SQL generation}
