% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-generation.R
\name{sqlDeleteAllRecords}
\alias{sqlDeleteAllRecords}
\title{generate sql deleting all records in a table}
\usage{
sqlDeleteAllRecords(conn, name, ...)
}
\arguments{
\item{conn}{a DBIConnector Object}

\item{name}{table in database to which primary key will be added}
}
\description{
`sqlDeleteAllRecords()` composes a single SQL statement to delete all records in an existing table
}
\seealso{
Other SQL generation: \code{\link{sqlAddIndex}},
  \code{\link{sqlAlterTableWithPrimaryKey}},
  \code{\link{sqlOnClause}},
  \code{\link{sqlReplaceInTable}},
  \code{\link{sqlSetClause}}, \code{\link{sqlUpdateTable}}
}
\concept{SQL generation}
