% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-generation.R
\name{sqlUpdateTable}
\alias{sqlUpdateTable}
\title{Compose an on clause for a table join}
\usage{
sqlUpdateTable(con, x, y, set, by, ...)
}
\arguments{
\item{x, y}{names of tbls to join. These will be quoted with dbQuoteIdentifier}

\item{set}{a character vector of columns to match .}

\item{by}{a character vector of columns to join by.}
}
\description{
`sqlUpdateTable()` generates a single SQL string that forms the an UPDATE statement
for UPDATE join between two tables.
}
\details{
The default method is MariaDB, MySQL compliant.
These methods are mostly useful for backend implementers.
}
\examples{

}
\seealso{
Other SQL generation: \code{\link{sqlAddIndex}},
  \code{\link{sqlAlterTableWithPrimaryKey}},
  \code{\link{sqlOnClause}},
  \code{\link{sqlReplaceInTable}},
  \code{\link{sqlSetClause}}
}
\concept{SQL generation}
