% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-update.R
\name{dbUpdateTable}
\alias{dbUpdateTable}
\title{update records in a database table}
\usage{
dbUpdateTable(con, target.table, data.update, set, by, ...)
}
\arguments{
\item{con}{an S4 object that inherits from DBIConnection. This object is used to communicate
with the database engine}

\item{target.table}{A character string specifying the unquoted DBMS table name, or the result
of a call to dbQuoteIdentifier().}

\item{data.update}{a data.frame (or coercible to data.frame).}

\item{set}{a named character vector of varibles}

\item{by}{a character vector of variables to join. If NULL, the default, *_join() will
          do a natural join, using all variables with common names across the two tables.
         A message lists the variables so that you can check they're right (to suppress
         the message, simply explicitly list the variables that you want to join).

To join by different variables on x and y use a named vector. For example, by = c("a" = "b") will match x.a to y.b.}
}
\description{
records in the target table in the database will be updated by the data provided in
the `data.update` data-frame. The default method provided, creates and populates a
temporary table using the data provided the `data.update` dataframe
}
\details{
An UPDATE statement is then constructed by the dbUpdateTable() method. The generated
 sql is then passed to dbExecute().
}
\examples{

}
\seealso{
Other DBIConnection generics: \code{\link{dbReplaceInTable}}
}
\concept{DBIConnection generics}
